debug

+++ cambrian/train/train_fsdp_multi.py
@@ -1487,7 +1487,7 @@ def prepare_multimodal_data(input_ids, labels, attention_mask, max_num_image_cro
 	attention_mask = torch.cat([attention_mask_image_full, attention_mask_newline_full, attention_mask_text], 1)
 	position_ids = torch.cat([position_ids_image_full, position_ids_newline_full, position_ids_text], 1)
 
-	dtype = torch.float32
+	dtype = torch.bfloat16
 
 	# prepare the 4D attention masks for regular attention and compressv attention
